\encoding{UTF-8}
\name{code}
\docType{data}
\alias{code}
\title{
 Administrative area code
}
\description{
  The administrative area code in China updated to July 2016.
}
\usage{code}

\format{A data.frame containing 3507 observations of 2 variables.}

\source{
  National Bureau of Statistics of the People's Republic of China.
}

\references{
  National Bureau of Statistics of the People's Republic of China.
}

\examples{
#load the dataset
data(code)
}

\keyword{datasets}
